Refactor: split parser and data exporter into focused modules
Split parser.py and data_exporter.py into smaller, focused modules
while preserving backward compatibility via re-exports:

Parser:
- parser_cache.py: cache key generation and persistence
- parser_common.py: shared XML helpers, coordinate parsing, metadata
- parser_standard.py: standard KML coordinate processing
- parser_gx_track.py: gx:Track format processing

Data exporter:
- export_pipeline.py: year/path/segment processing
- export_writers.py: airport and metadata export
- export_reconciler.py: statistics reconciliation

All existing imports continue to work unchanged.
