+++ kml_heatmap/data_exporter.py
@@ -24,352 +24,14 @@ This approach ensures:
 import os
 import json
 import shutil
-from bisect import bisect_left, bisect_right
 from concurrent.futures import ThreadPoolExecutor, as_completed
-from typing import List, Dict, Any, Optional, Tuple
+from typing import List, Dict, Any
 
-from .airports import extract_airport_name
 from .logger import logger
-from .constants import (
-    DATA_RESOLUTION,
-    HEATMAP_GRADIENT,
-    KM_TO_NAUTICAL_MILES,
-    METERS_TO_FEET,
-    SPEED_WINDOW_SECONDS,
-    CRUISE_ALTITUDE_THRESHOLD_FT,
-)
-from .geometry import haversine_distance
-from .helpers import parse_iso_timestamp, calculate_duration_seconds, format_flight_time
-from .segment_calculator import (
-    extract_segment_speeds,
-    build_time_indexed_segments,
-    calculate_windowed_groundspeed,
-    calculate_fallback_groundspeed,
-    calculate_path_distance,
-    update_cruise_statistics,
-)
-
-
-def export_airports_data(
-    unique_airports: List[Dict[str, Any]],
-    output_dir: str,
-    strip_timestamps: bool = False,
-) -> Tuple[str, int]:
-    """
-    Export airport data to JSON file.
-
-    Args:
-        unique_airports: List of airport dictionaries
-        output_dir: Output directory
-        strip_timestamps: Whether to strip timestamps for privacy
-
-    Returns:
-        Tuple of (output_file_path, file_size_bytes)
-    """
-    valid_airports = []
-    seen_locations = set()
-
-    for apt in unique_airports:
-        full_name = apt.get("name", "Unknown")
-        is_at_path_end = apt.get("is_at_path_end", False)
-        airport_name = extract_airport_name(full_name, is_at_path_end)
-
-        if not airport_name:
-            continue
-
-        location_key = f"{apt['lat']:.4f},{apt['lon']:.4f}"
-
-        if location_key in seen_locations:
-            continue
-
-        seen_locations.add(location_key)
-
-        airport_data = {
-            "lat": apt["lat"],
-            "lon": apt["lon"],
-            "name": airport_name,
-            "flight_count": len(apt["timestamps"]) if apt["timestamps"] else 1,
-        }
-
-        if not strip_timestamps:
-            airport_data["timestamps"] = apt["timestamps"]
-
-        valid_airports.append(airport_data)
-
-    airports_data = {"airports": valid_airports}
-    airports_file = os.path.join(output_dir, "airports.js")
-
-    with open(airports_file, "w") as f:
-        f.write("window.KML_AIRPORTS = ")
-        json.dump(airports_data, f, separators=(",", ":"), sort_keys=True)
-        f.write(";")
-
-    file_size = os.path.getsize(airports_file)
-
-    logger.info(
-        f"  ✓ Airports: {len(valid_airports)} locations ({file_size / 1024:.1f} KB)"
-    )
-
-    return airports_file, file_size
-
-
-def export_metadata(
-    stats: Dict[str, Any],
-    min_alt_m: float,
-    max_alt_m: float,
-    min_groundspeed_knots: float,
-    max_groundspeed_knots: float,
-    available_years: List[int],
-    output_dir: str,
-    file_structure: Optional[Dict[str, Any]] = None,
-) -> Tuple[str, int]:
-    """
-    Export metadata including statistics and ranges.
-
-    Args:
-        stats: Statistics dictionary
-        min_alt_m: Minimum altitude in meters
-        max_alt_m: Maximum altitude in meters
-        min_groundspeed_knots: Minimum groundspeed
-        max_groundspeed_knots: Maximum groundspeed
-        available_years: List of available years
-        output_dir: Output directory
-        file_structure: File structure mapping (year -> resolution list)
-
-    Returns:
-        Tuple of (output_file_path, file_size_bytes)
-    """
-    # Handle infinity case
-    if min_groundspeed_knots == float("inf"):
-        min_groundspeed_knots = 0.0
-
-    meta_data = {
-        "stats": stats,
-        "min_alt_m": min_alt_m,
-        "max_alt_m": max_alt_m,
-        "min_groundspeed_knots": round(min_groundspeed_knots, 1),
-        "max_groundspeed_knots": round(max_groundspeed_knots, 1),
-        "gradient": HEATMAP_GRADIENT,
-        "available_years": available_years,
-    }
-
-    if file_structure is not None:
-        meta_data["file_structure"] = file_structure
-
-    meta_file = os.path.join(output_dir, "metadata.js")
-
-    with open(meta_file, "w") as f:
-        f.write("window.KML_METADATA = ")
-        json.dump(meta_data, f, separators=(",", ":"), sort_keys=True)
-        f.write(";")
-
-    file_size = os.path.getsize(meta_file)
-
-    logger.info(f"  ✓ Metadata: {file_size / 1024:.1f} KB")
-
-    return meta_file, file_size
-
-
-def collect_unique_years(all_path_metadata: List[Dict[str, Any]]) -> List[int]:
-    """
-    Collect unique years from path metadata.
-
-    Args:
-        all_path_metadata: List of path metadata dictionaries
-
-    Returns:
-        Sorted list of unique years
-    """
-    unique_years = set()
-
-    for meta in all_path_metadata:
-        year = meta.get("year")
-        if year:
-            unique_years.add(year)
-
-    return sorted(list(unique_years))
-
-
-def _build_path_info(
-    path: List[List[Any]],
-    metadata: Dict[str, Any],
-    local_idx: int,
-    path_year: Any,
-) -> Tuple[Dict[str, Any], float, float, float]:
-    """Build path info entry and compute path metrics.
-
-    Args:
-        path: Path coordinate list
-        metadata: Path metadata dictionary
-        local_idx: Local path index within the year
-        path_year: Year from metadata
-
-    Returns:
-        Tuple of (info_dict, path_duration_seconds, path_distance_km, path_distance_nm)
-    """
-    airport_name = metadata.get("airport_name", "")
-    start_airport = None
-    end_airport = None
-
-    if airport_name and " - " in airport_name:
-        parts = airport_name.split(" - ")
-        if len(parts) == 2:
-            start_airport = parts[0].strip()
-            end_airport = parts[1].strip()
-
-    path_duration_seconds = 0.0
-    start_ts = metadata.get("timestamp")
-    end_ts = metadata.get("end_timestamp")
-
-    if start_ts and end_ts:
-        path_duration_seconds = calculate_duration_seconds(start_ts, end_ts)
-        if path_duration_seconds == 0:
-            logger.debug(f"  Could not parse timestamps '{start_ts}' -> '{end_ts}'")
-
-    path_distance_km = calculate_path_distance(path)
-    path_distance_nm = path_distance_km * KM_TO_NAUTICAL_MILES
-
-    info: Dict[str, Any] = {
-        "id": local_idx,
-        "start_airport": start_airport,
-        "end_airport": end_airport,
-        "start_coords": [path[0][0], path[0][1]],
-        "end_coords": [path[-1][0], path[-1][1]],
-        "segment_count": len(path) - 1,
-        "year": path_year,
-    }
-    if "aircraft_registration" in metadata:
-        info["aircraft_registration"] = metadata["aircraft_registration"]
-    if "aircraft_type" in metadata:
-        info["aircraft_type"] = metadata["aircraft_type"]
-
-    return info, path_duration_seconds, path_distance_km, path_distance_nm
-
-
-def _process_path_segments(
-    path: List[List[Any]],
-    local_idx: int,
-    path_distance_km: float,
-    path_duration_seconds: float,
-) -> Tuple[List[Dict[str, Any]], float, float, float, float, Dict[int, float]]:
-    """Process path segments calculating groundspeed and cruise statistics.
-
-    Args:
-        path: Path coordinate list
-        local_idx: Local path index
-        path_distance_km: Total path distance in km
-        path_duration_seconds: Total path duration in seconds
-
-    Returns:
-        Tuple of (segments, max_groundspeed, min_groundspeed,
-                  cruise_distance, cruise_time, cruise_altitude_histogram)
-    """
-    segments: List[Dict[str, Any]] = []
-    max_groundspeed = 0.0
-    min_groundspeed = float("inf")
-    cruise_distance = 0.0
-    cruise_time = 0.0
-    cruise_altitude_histogram: Dict[int, float] = {}
-
-    ground_level_m = min(coord[2] for coord in path) if path else 0
-
-    path_start_time = None
-    for coord in path:
-        if len(coord) >= 4:
-            path_start_time = parse_iso_timestamp(coord[3])
-            if path_start_time:
-                break
-
-    segment_speeds = extract_segment_speeds(path, path_start_time)
-    timestamp_list, time_indexed_segments = build_time_indexed_segments(segment_speeds)
-
-    for i in range(len(path) - 1):
-        coord1 = path[i]
-        coord2 = path[i + 1]
-        lat1, lon1, alt1_m = coord1[0], coord1[1], coord1[2]
-        lat2, lon2, alt2_m = coord2[0], coord2[1], coord2[2]
-
-        avg_alt_m = (alt1_m + alt2_m) / 2
-        avg_alt_ft = round(avg_alt_m * METERS_TO_FEET / 100) * 100
-
-        groundspeed_knots = 0.0
-        current_segment = segment_speeds[i]
-        current_timestamp = current_segment["timestamp"]
-        current_relative_time = current_segment["relative_time"]
-
-        if current_timestamp is not None and timestamp_list:
-            groundspeed_knots = calculate_windowed_groundspeed(
-                current_timestamp, timestamp_list, time_indexed_segments
-            )
-
-        if groundspeed_knots == 0:
-            segment_distance_km = haversine_distance(lat1, lon1, lat2, lon2)
-            groundspeed_knots = calculate_fallback_groundspeed(
-                segment_distance_km, path_distance_km, path_duration_seconds
-            )
-
-        if groundspeed_knots > 0:
-            if groundspeed_knots > max_groundspeed:
-                max_groundspeed = groundspeed_knots
-            if groundspeed_knots < min_groundspeed:
-                min_groundspeed = groundspeed_knots
-
-            altitude_agl_m = avg_alt_m - ground_level_m
-            altitude_agl_ft = altitude_agl_m * METERS_TO_FEET
-            if altitude_agl_ft > CRUISE_ALTITUDE_THRESHOLD_FT:
-                if current_timestamp is not None and timestamp_list:
-                    current_ts = current_timestamp.timestamp()
-                    half_window = SPEED_WINDOW_SECONDS / 2
-                    w_start = bisect_left(timestamp_list, current_ts - half_window)
-                    w_end = bisect_right(timestamp_list, current_ts + half_window)
-                    w_distance = sum(
-                        time_indexed_segments[j]["distance"]
-                        for j in range(w_start, w_end)
-                    )
-                    w_time = sum(
-                        time_indexed_segments[j]["time_delta"]
-                        for j in range(w_start, w_end)
-                    )
-                else:
-                    w_distance = 0.0
-                    w_time = 0.0
-
-                cruise_stats: Dict[str, Any] = {
-                    "total_distance": 0.0,
-                    "total_time": 0.0,
-                    "altitude_histogram": {},
-                }
-                update_cruise_statistics(
-                    altitude_agl_ft, w_time, w_distance, cruise_stats
-                )
-                cruise_distance += float(cruise_stats["total_distance"])
-                cruise_time += float(cruise_stats["total_time"])
-                hist: Dict[int, float] = cruise_stats["altitude_histogram"]
-                for alt_bin, time_spent in hist.items():
-                    if alt_bin not in cruise_altitude_histogram:
-                        cruise_altitude_histogram[alt_bin] = 0.0
-                    cruise_altitude_histogram[alt_bin] += time_spent
-
-        if lat1 != lat2 or lon1 != lon2:
-            segment_data: Dict[str, Any] = {
-                "coords": [[lat1, lon1], [lat2, lon2]],
-                "altitude_ft": avg_alt_ft,
-                "altitude_m": round(avg_alt_m, 0),
-                "groundspeed_knots": round(groundspeed_knots, 1),
-                "path_id": local_idx,
-            }
-            if current_relative_time is not None:
-                segment_data["time"] = round(current_relative_time, 1)
-            segments.append(segment_data)
-
-    return (
-        segments,
-        max_groundspeed,
-        min_groundspeed,
-        cruise_distance,
-        cruise_time,
-        cruise_altitude_histogram,
-    )
+from .constants import DATA_RESOLUTION
+from .export_writers import export_airports_data, export_metadata, collect_unique_years
+from .export_reconciler import _recalculate_stats_from_segments
+from .export_pipeline import _build_path_info, _process_path_segments
 
 
 def process_year_data(
@@ -539,138 +201,6 @@ def process_year_data(
         "full_res_segments": year_full_res_segments,
         "full_res_path_info": year_full_res_path_info,
     }
-
-
-def _recalculate_stats_from_segments(
-    stats: Dict[str, Any],
-    segments: List[Dict[str, Any]],
-    path_info_list: List[Dict[str, Any]],
-) -> None:
-    """Recalculate statistics from exported segment data (authoritative source).
-
-    The initial stats from statistics.py are computed from raw parsed data, but
-    the frontend only sees the exported segments. This function overwrites stats
-    to match the segment data exactly, ensuring the statistics panel is consistent
-    with what the frontend can independently verify. Fields overwritten:
-    total_points, altitude ranges, altitude gain, groundspeed stats, cruise
-    altitude, flight time, and per-aircraft times/distances.
-
-    Args:
-        stats: Statistics dictionary (modified in place)
-        segments: Full resolution path segments
-        path_info_list: Full resolution path info entries
-    """
-    stats["total_points"] = len(segments) * 2
-
-    if not segments:
-        return
-
-    altitudes_m = [seg.get("altitude_m", 0) for seg in segments]
-    min_alt_m = min(altitudes_m)
-    max_alt_m = max(altitudes_m)
-    stats["min_altitude_m"] = min_alt_m
-    stats["max_altitude_m"] = max_alt_m
-    stats["min_altitude_ft"] = min_alt_m * METERS_TO_FEET
-    stats["max_altitude_ft"] = max_alt_m * METERS_TO_FEET
-
-    cruise_threshold = min_alt_m * METERS_TO_FEET + 1000
-
-    total_gain_m = 0.0
-    prev_alt = None
-    groundspeed_sum = 0.0
-    groundspeed_count = 0
-    cruise_speed_sum = 0.0
-    cruise_speed_count = 0
-    altitude_bins: Dict[int, int] = {}
-    path_durations: Dict[int, List[float]] = {}
-
-    for seg in segments:
-        alt_m = seg.get("altitude_m", 0)
-        alt_ft = seg.get("altitude_ft", 0)
-        gs = seg.get("groundspeed_knots", 0)
-
-        if prev_alt is not None and alt_m > prev_alt:
-            total_gain_m += alt_m - prev_alt
-        prev_alt = alt_m
-
-        if gs > 0:
-            groundspeed_sum += gs
-            groundspeed_count += 1
-
-        if alt_ft > cruise_threshold:
-            if gs > 0:
-                cruise_speed_sum += gs
-                cruise_speed_count += 1
-            if "time" in seg:
-                bin_alt = round(alt_ft / 100) * 100
-                altitude_bins[bin_alt] = altitude_bins.get(bin_alt, 0) + 1
-
-        if "time" in seg and "path_id" in seg:
-            path_id = seg["path_id"]
-            if path_id not in path_durations:
-                path_durations[path_id] = []
-            path_durations[path_id].append(seg["time"])
-
-    stats["total_altitude_gain_m"] = total_gain_m
-    stats["total_altitude_gain_ft"] = total_gain_m * METERS_TO_FEET
-    stats["average_groundspeed_knots"] = (
-        groundspeed_sum / groundspeed_count if groundspeed_count > 0 else 0
-    )
-    stats["cruise_speed_knots"] = (
-        cruise_speed_sum / cruise_speed_count if cruise_speed_count > 0 else 0
-    )
-
-    if altitude_bins:
-        stats["most_common_cruise_altitude_ft"] = max(
-            altitude_bins.keys(), key=lambda k: altitude_bins[k]
-        )
-        stats["most_common_cruise_altitude_m"] = round(
-            stats["most_common_cruise_altitude_ft"] * 0.3048
-        )
-
-    total_flight_time = 0.0
-    for times in path_durations.values():
-        if len(times) >= 2:
-            total_flight_time += max(times) - min(times)
-    stats["total_flight_time_seconds"] = total_flight_time
-
-    if stats.get("aircraft_list"):
-        aircraft_times: Dict[str, float] = {}
-        aircraft_distances: Dict[str, float] = {}
-
-        for pi in path_info_list:
-            reg = pi.get("aircraft_registration")
-            path_id = pi.get("id")
-            if reg and path_id is not None and path_id in path_durations:
-                if reg not in aircraft_times:
-                    aircraft_times[reg] = 0
-                    aircraft_distances[reg] = 0.0
-                times = path_durations[path_id]
-                if len(times) >= 2:
-                    aircraft_times[reg] += max(times) - min(times)
-
-        for segment in segments:
-            path_id = segment.get("path_id")
-            if path_id is not None and path_id < len(path_info_list):
-                pi = path_info_list[path_id]
-                reg = pi.get("aircraft_registration")
-                if reg and reg in aircraft_distances:
-                    coords = segment.get("coords", [])
-                    if len(coords) == 2:
-                        lat1, lon1 = coords[0]
-                        lat2, lon2 = coords[1]
-                        aircraft_distances[reg] += haversine_distance(
-                            lat1, lon1, lat2, lon2
-                        )
-
-        for aircraft in stats["aircraft_list"]:
-            reg = aircraft["registration"]
-            if reg in aircraft_times:
-                flight_time_seconds = aircraft_times[reg]
-                flight_distance_km = aircraft_distances.get(reg, 0.0)
-                aircraft["flight_time_seconds"] = flight_time_seconds
-                aircraft["flight_distance_km"] = flight_distance_km
-                aircraft["flight_time_str"] = format_flight_time(flight_time_seconds)
 
 
 def export_all_data(
