+++ kml_heatmap/parser.py
@@ -30,37 +30,35 @@ Example:
     Loaded 1234 points from 3 paths
 """
 
-import json
-import re
-from datetime import datetime
 from pathlib import Path
 from typing import Any, Dict, List, Optional, Tuple
 
 from lxml import etree as ET
 
-from .aircraft import parse_aircraft_from_filename
-from .cache import CACHE_DIR, atomic_json_write
-from .kml_parsers import validate_and_normalize_coordinate
 from .logger import logger
 from .airport_lookup import (
     extract_icao_codes_from_name,
     standardize_airport_name,
 )
-from .constants import (
-    MID_FLIGHT_MIN_ALTITUDE_M,
-    MID_FLIGHT_MAX_VARIATION_M,
-    LANDING_MAX_VARIATION_M,
-    LANDING_MAX_ALTITUDE_M,
-    LANDING_FALLBACK_ALTITUDE_M,
-    PATH_SAMPLE_MAX_SIZE,
-    PATH_SAMPLE_MIN_SIZE,
-    LAT_MIN,
-    LAT_MAX,
-    LON_MIN,
-    LON_MAX,
-    ALT_MIN_M,
-    ALT_MAX_M,
+from .parser_cache import KML_CACHE_DIR  # noqa: F401 - re-exported for test compatibility
+from .parser_cache import (
+    get_cache_key as _get_cache_key,
+    load_cached_parse,
+    save_to_cache as _save_to_cache,
 )
+from .parser_common import (
+    extract_year_from_timestamp,
+    sample_path_altitudes,
+    is_mid_flight_start,
+    is_valid_landing,
+    parse_coordinate_point,
+    find_xml_element,
+    find_xml_elements,
+    extract_charterware_timestamp,  # noqa: F401 - re-exported for backward compatibility
+    extract_placemark_metadata,
+)
+from .parser_standard import process_standard_coordinates
+from .parser_gx_track import process_gx_track
 
 __all__ = [
     "get_cache_key",
@@ -81,75 +79,10 @@ __all__ = [
     "process_gx_track",
 ]
 
-# Pre-compiled regex patterns for performance
-DATE_PATTERN = re.compile(r"(\d{2}\s+\w{3}\s+\d{4}|\d{4}-\d{2}-\d{2})")
-YEAR_PATTERN = re.compile(r"\b(20\d{2})\b")
-
-# Pre-computed validation ranges (avoid function call overhead)
-LAT_RANGE = (LAT_MIN, LAT_MAX)
-LON_RANGE = (LON_MIN, LON_MAX)
-ALT_RANGE = (ALT_MIN_M, ALT_MAX_M)
-
-# KML parse cache subdirectory
-KML_CACHE_DIR = CACHE_DIR / "kml"
-
 
 def get_cache_key(kml_file: str) -> Tuple[Optional[Path], bool]:
-    """Generate cache key based on file path and modification time.
-
-    Args:
-        kml_file: Path to KML file
-
-    Returns:
-        Tuple of (cache_path, is_valid) where is_valid indicates if cached data can be used
-    """
-    kml_path = Path(kml_file)
-
-    # Create cache directory if it doesn't exist
-    KML_CACHE_DIR.mkdir(parents=True, exist_ok=True)
-
-    # Get file modification time
-    try:
-        mtime = kml_path.stat().st_mtime
-    except (OSError, FileNotFoundError):
-        return None, False
-
-    # Create cache filename from KML filename and modification time
-    cache_name = f"{kml_path.stem}_{int(mtime)}.json"
-    cache_path = KML_CACHE_DIR / cache_name
-
-    # Check if cache file exists
-    if cache_path.exists():
-        return cache_path, True
-
-    # Clean up old cache files for this KML file (different mtime)
-    for old_cache in KML_CACHE_DIR.glob(f"{kml_path.stem}_*.json"):
-        if old_cache != cache_path:  # Don't delete the one we're about to write
-            try:
-                old_cache.unlink()
-            except OSError:
-                pass
-
-    return cache_path, False
-
-
-def load_cached_parse(
-    cache_path: Path,
-) -> Optional[Tuple[List[List[float]], List[List[List[float]]], List[Dict[str, Any]]]]:
-    """Load cached parse results.
-
-    Args:
-        cache_path: Path to cache file
-
-    Returns:
-        Tuple of (coordinates, path_groups, path_metadata) or None if cache invalid
-    """
-    try:
-        with open(cache_path, "r") as f:
-            cached = json.load(f)
-        return cached["coordinates"], cached["path_groups"], cached["path_metadata"]
-    except (json.JSONDecodeError, KeyError, OSError):
-        return None
+    """Generate cache key using the module-level KML_CACHE_DIR."""
+    return _get_cache_key(kml_file, cache_dir=KML_CACHE_DIR)
 
 
 def save_to_cache(
@@ -158,662 +91,10 @@ def save_to_cache(
     path_groups: List[List[List[float]]],
     path_metadata: List[Dict[str, Any]],
 ) -> None:
-    """Save parse results to cache.
-
-    Args:
-        cache_path: Path to cache file
-        coordinates: List of coordinates
-        path_groups: List of path groups
-        path_metadata: List of path metadata dicts
-    """
-    atomic_json_write(
-        cache_path,
-        {
-            "coordinates": coordinates,
-            "path_groups": path_groups,
-            "path_metadata": path_metadata,
-        },
-        KML_CACHE_DIR,
+    """Save parse results using the module-level KML_CACHE_DIR."""
+    _save_to_cache(
+        cache_path, coordinates, path_groups, path_metadata, cache_dir=KML_CACHE_DIR
     )
-
-
-def extract_year_from_timestamp(timestamp: Optional[str]) -> Optional[int]:
-    """Extract year from timestamp string.
-
-    Args:
-        timestamp: Timestamp string in ISO format (e.g., "2025-03-03T08:58:01Z")
-                   or other date formats
-
-    Returns:
-        Year as integer, or None if extraction fails
-    """
-    if not timestamp:
-        return None
-
-    try:
-        # Try to parse ISO format timestamp (e.g., "2025-03-03T08:58:01Z")
-        if "T" in timestamp:
-            dt = datetime.fromisoformat(timestamp.replace("Z", "+00:00"))
-            return dt.year
-        # Try to extract year from date string (e.g., "03 Mar 2025" or "2025-03-03")
-        year_match = YEAR_PATTERN.search(timestamp)
-        if year_match:
-            return int(year_match.group(1))
-    except (ValueError, AttributeError, TypeError) as e:
-        logger.debug(f"Could not parse timestamp '{timestamp}': {e}")
-
-    return None
-
-
-def sample_path_altitudes(
-    path: List[List[float]], from_end: bool = False
-) -> Optional[Dict[str, float]]:
-    """
-    Extract altitude statistics from a path sample.
-
-    Args:
-        path: List of [lat, lon, alt] coordinates
-        from_end: If True, sample from end of path; otherwise from start
-
-    Returns:
-        Dict with 'min', 'max', 'variation' keys, or None if sample too small
-    """
-    if len(path) <= 10:
-        return None
-
-    sample_size = min(PATH_SAMPLE_MAX_SIZE, len(path) // 4)
-    if sample_size <= PATH_SAMPLE_MIN_SIZE:
-        return None
-
-    sample = path[-sample_size:] if from_end else path[:sample_size]
-    alts = [coord[2] for coord in sample]
-    return {"min": min(alts), "max": max(alts), "variation": max(alts) - min(alts)}
-
-
-def is_mid_flight_start(
-    path: List[List[float]], start_alt: float, debug: bool = False
-) -> bool:
-    """
-    Detect if a path started mid-flight by analyzing altitude patterns.
-
-    A mid-flight start is characterized by:
-    1. Starting at altitude (> MID_FLIGHT_MIN_ALTITUDE) AND
-    2. NOT descending/ascending significantly in the first part of the path
-
-    Args:
-        path: List of [lat, lon, alt] coordinates
-        start_alt: Starting altitude in meters
-        debug: Enable debug output
-
-    Returns:
-        bool: True if this is a mid-flight start
-    """
-    sample = sample_path_altitudes(path, from_end=False)
-    if not sample:
-        return False
-
-    # Mid-flight indicators:
-    # - Starting altitude above typical airports
-    # - AND altitude variation in first part is small (not climbing/descending much)
-    is_mid_flight = (
-        start_alt > MID_FLIGHT_MIN_ALTITUDE_M
-        and sample["variation"] < MID_FLIGHT_MAX_VARIATION_M
-    )
-
-    if is_mid_flight and debug:
-        logger.debug(
-            f"Detected mid-flight start at {start_alt:.0f}m (variation: {sample['variation']:.0f}m)"
-        )
-
-    return is_mid_flight
-
-
-def is_valid_landing(
-    path: List[List[float]], end_alt: float, debug: bool = False
-) -> bool:
-    """
-    Check if a path ends with a valid landing.
-
-    A valid landing shows descent or stable low altitude at the end.
-
-    Args:
-        path: List of [lat, lon, alt] coordinates
-        end_alt: Ending altitude in meters
-        debug: Enable debug output
-
-    Returns:
-        bool: True if this is a valid landing
-    """
-    sample = sample_path_altitudes(path, from_end=True)
-    if not sample:
-        # Short path, just accept if altitude seems reasonable
-        return end_alt < LANDING_FALLBACK_ALTITUDE_M
-
-    # Valid landing: either descending significantly OR stable at low variation
-    # Also accept any endpoint if variation at end is small - indicates stable landing
-    return (
-        sample["variation"] < LANDING_MAX_VARIATION_M
-        or end_alt < LANDING_MAX_ALTITUDE_M
-    )
-
-
-def parse_coordinate_point(
-    point: str, kml_file: str
-) -> Optional[Tuple[float, float, Optional[float]]]:
-    """
-    Parse a single coordinate point from KML format.
-
-    Args:
-        point: Coordinate string in format "lon,lat,alt" or "lon,lat"
-        kml_file: Path to KML file (for error messages)
-
-    Returns:
-        Tuple of (lat, lon, alt) or None if invalid
-    """
-    point = point.strip()
-    if not point:
-        return None
-
-    parts = point.split(",")
-    if len(parts) < 2:
-        return None
-
-    try:
-        lon = float(parts[0])
-        lat = float(parts[1])
-        alt = float(parts[2]) if len(parts) >= 3 else None
-
-        # Use centralized validation and normalization
-        return validate_and_normalize_coordinate(lat, lon, alt, Path(kml_file).name)
-    except ValueError as e:
-        logger.debug(f"Failed to parse coordinate '{point}': {e}")
-        return None
-
-
-def find_xml_element(
-    parent: Any, namespaced_path: str, fallback_path: str, namespaces: Dict[str, str]
-) -> Optional[Any]:
-    """
-    Find XML element trying namespaced path first, then fallback without namespace.
-
-    Args:
-        parent: Parent XML element to search within
-        namespaced_path: XPath with namespace prefix (e.g., './/kml:name')
-        fallback_path: XPath without namespace (e.g., './/name')
-        namespaces: XML namespace dict
-
-    Returns:
-        Found XML element or None
-    """
-    elem = parent.find(namespaced_path, namespaces)
-    if elem is None:
-        elem = parent.find(fallback_path)
-    return elem
-
-
-def find_xml_elements(
-    parent: Any, namespaced_path: str, fallback_path: str, namespaces: Dict[str, str]
-) -> List[Any]:
-    """
-    Find XML elements trying namespaced path first, then fallback without namespace.
-
-    Args:
-        parent: Parent XML element to search within
-        namespaced_path: XPath with namespace prefix (e.g., './/kml:when')
-        fallback_path: XPath without namespace (e.g., './/when')
-        namespaces: XML namespace dict
-
-    Returns:
-        List of found XML elements (empty list if none found)
-    """
-    elems: List[Any] = parent.findall(namespaced_path, namespaces)
-    if not elems:
-        elems = parent.findall(fallback_path)
-    return elems
-
-
-def extract_charterware_timestamp(description: str) -> Optional[str]:
-    """
-    Extract timestamp from Charterware description field.
-
-    Example: "Flight Jan 12 2026 03:01PM path of OE-AKI"
-    Returns ISO-format timestamp like "2026-01-12T15:01:00Z"
-
-    Args:
-        description: Description text from Charterware KML
-
-    Returns:
-        ISO-format timestamp or None if extraction fails
-    """
-    if not description:
-        return None
-
-    # Pattern: "Flight Jan 12 2026 03:01PM" or "Flight January 12 2026 03:01PM"
-    pattern = r"Flight\s+(\w{3,9})\s+(\d{1,2})\s+(\d{4})\s+(\d{2}):(\d{2})(AM|PM)"
-    match = re.search(pattern, description)
-
-    if match:
-        month_str, day, year, hour, minute, meridiem = match.groups()
-
-        # Convert 12-hour to 24-hour
-        hour = int(hour)
-        if meridiem == "PM" and hour != 12:
-            hour += 12
-        elif meridiem == "AM" and hour == 12:
-            hour = 0
-
-        # Parse month name (supports both short and full month names)
-        try:
-            dt_str = f"{day} {month_str} {year} {hour:02d}:{minute}"
-            # Try short month name first (Jan, Feb, etc.)
-            dt = datetime.strptime(dt_str, "%d %b %Y %H:%M")
-        except ValueError:
-            # Try full month name (January, February, etc.)
-            try:
-                dt = datetime.strptime(dt_str, "%d %B %Y %H:%M")
-            except ValueError:
-                logger.debug(f"Failed to parse Charterware timestamp: {description}")
-                return None
-
-        # Return ISO format with UTC timezone
-        return dt.isoformat() + "Z"
-
-    return None
-
-
-def extract_placemark_metadata(
-    placemark: Any, namespaces: Dict[str, str], kml_file: Optional[str] = None
-) -> Dict[str, Any]:
-    """
-    Extract metadata from a KML Placemark element.
-
-    Args:
-        placemark: XML element representing a Placemark
-        namespaces: XML namespace dict
-        kml_file: Optional KML filename (unused, kept for backward compatibility)
-
-    Returns:
-        Dict with 'airport_name', 'timestamp', 'end_timestamp', 'year' keys
-    """
-    # Extract name from KML
-    name_elem = find_xml_element(placemark, ".//kml:name", ".//name", namespaces)
-    kml_name = (
-        name_elem.text.strip() if name_elem is not None and name_elem.text else None
-    )
-
-    # Standardize airport name using ICAO codes from the name itself
-    airport_name = standardize_airport_name(kml_name)
-
-    # Extract timestamps - both start and end for tracks with multiple when elements
-    time_elems = find_xml_elements(placemark, ".//kml:when", ".//when", namespaces)
-
-    # Also try TimeStamp element (single timestamp)
-    if not time_elems:
-        time_elem = find_xml_element(
-            placemark, ".//kml:TimeStamp/kml:when", ".//TimeStamp/when", namespaces
-        )
-        if time_elem is not None:
-            time_elems = [time_elem]
-
-    timestamp = None
-    end_timestamp = None
-    if time_elems and len(time_elems) > 0:
-        if time_elems[0].text:
-            timestamp = time_elems[0].text.strip()
-        # Get last timestamp if multiple exist
-        if len(time_elems) > 1 and time_elems[-1].text:
-            end_timestamp = time_elems[-1].text.strip()
-    elif kml_name:
-        # Try to extract date from original KML name (before standardization)
-        # (e.g., "Log Start: 03 Mar 2025 08:58 Z" or "EDDS to EDDP - 16 Aug 2026")
-        match = DATE_PATTERN.search(kml_name)
-        if match:
-            timestamp = match.group(1)
-
-    # If still no timestamp, check description for Charterware format
-    if timestamp is None:
-        desc_elem = find_xml_element(
-            placemark, ".//kml:description", ".//description", namespaces
-        )
-        if desc_elem is not None and desc_elem.text:
-            # Try Charterware format: "Flight Jan 12 2026 03:01PM path of OE-AKI"
-            timestamp = extract_charterware_timestamp(desc_elem.text.strip())
-
-    year = extract_year_from_timestamp(timestamp)
-
-    return {
-        "airport_name": airport_name,
-        "timestamp": timestamp,
-        "end_timestamp": end_timestamp,
-        "year": year,
-    }
-
-
-def _build_path_metadata_dict(
-    kml_file: str,
-    path_start: List[float],
-    airport_name: Optional[str],
-    timestamp: Optional[str],
-    end_timestamp: Optional[str],
-) -> Dict[str, Any]:
-    """
-    Build path metadata dictionary.
-
-    Args:
-        kml_file: Path to KML file
-        path_start: First coordinate of path [lat, lon, alt]
-        airport_name: Airport name from metadata
-        timestamp: Start timestamp
-        end_timestamp: End timestamp
-
-    Returns:
-        Dict with path metadata
-    """
-    year = extract_year_from_timestamp(timestamp)
-    aircraft_info = parse_aircraft_from_filename(Path(kml_file).name)
-
-    # For Charterware files, use route information for airport name
-    # Route format: DEPARTURE-ARRIVAL (e.g., LOAV-LOAV or EDDF-EDDM)
-    # Convert to exporter format: "DEPARTURE - ARRIVAL" (with spaces around hyphen)
-    if (
-        aircraft_info
-        and aircraft_info.get("route")
-        and aircraft_info.get("format") == "charterware"
-    ):
-        route = aircraft_info.get("route")
-        # Extract airports from route and format for exporter
-        if route and "-" in route:
-            route_parts = route.split("-")
-            departure_airport = route_parts[0]
-            arrival_airport = route_parts[1]
-            # Use route as airport_name if current name is empty or is not a 4-letter ICAO code
-            # ICAO codes are exactly 4 uppercase letters (e.g., LOAV, EDDF, EDDM)
-            # Aircraft registrations contain hyphens (e.g., OE-AKI, D-EXYZ)
-            # Format: "DEPARTURE - ARRIVAL" (with spaces) for exporter compatibility
-            if not airport_name or len(airport_name) != 4:
-                airport_name = f"{departure_airport} - {arrival_airport}"
-                # Look up full airport names from ICAO codes
-                airport_name = standardize_airport_name(airport_name)
-
-    meta = {
-        "timestamp": timestamp,
-        "end_timestamp": end_timestamp,
-        "filename": Path(kml_file).name,
-        "start_point": path_start,
-        "airport_name": airport_name,
-        "year": year,
-    }
-
-    # Add aircraft info if available
-    if aircraft_info:
-        meta["aircraft_registration"] = aircraft_info.get("registration")
-
-        # Handle optional aircraft type (may be None for Charterware)
-        aircraft_type = aircraft_info.get("type")
-        if aircraft_type is not None:
-            meta["aircraft_type"] = aircraft_type
-
-        # Handle optional route (Charterware specific)
-        route = aircraft_info.get("route")
-        if route is not None:
-            meta["route"] = route
-
-    return meta
-
-
-def process_standard_coordinates(
-    coord_elements: List[Any],
-    coord_to_metadata: Dict[int, Dict[str, Any]],
-    kml_file: str,
-    coordinates: List[List[float]],
-    path_groups: List[List[List[float]]],
-    path_metadata: List[Dict[str, Any]],
-) -> None:
-    """
-    Process standard KML <coordinates> elements.
-
-    Args:
-        coord_elements: List of XML coordinate elements
-        coord_to_metadata: Mapping from element ID to metadata dict
-        kml_file: Path to KML file
-        coordinates: Output list for all [lat, lon] pairs
-        path_groups: Output list for path groups with altitude
-        path_metadata: Output list for path metadata
-    """
-    for idx, coord_elem in enumerate(coord_elements):
-        # Handle None text
-        if coord_elem.text is None:
-            logger.debug(f"Coordinate element {idx} has None text, skipping")
-            continue
-
-        coord_text = coord_elem.text.strip()
-        if not coord_text:
-            logger.debug(f"Coordinate element {idx} has empty text, skipping")
-            continue
-
-        # Get metadata for this coordinate element
-        metadata = coord_to_metadata.get(id(coord_elem), {})
-        airport_name = metadata.get("airport_name")
-        timestamp = metadata.get("timestamp")
-        end_timestamp = metadata.get("end_timestamp")
-
-        # Split by whitespace (spaces, tabs, newlines)
-        points = coord_text.split()
-
-        # Create a new path group for this coordinate element
-        current_path = []
-        element_coords = 0
-
-        for point in points:
-            parsed = parse_coordinate_point(point, kml_file)
-            if parsed is None:
-                continue
-
-            lat, lon, alt = parsed
-
-            # Swap to [lat, lon] for leaflet
-            coordinates.append([lat, lon])
-
-            # Add to current path group with altitude
-            if alt is not None:
-                current_path.append([lat, lon, alt])
-
-            element_coords += 1
-
-        # Add this path group to the list if it has coordinates
-        if current_path:
-            # Do NOT generate synthetic timestamps for Charterware files
-            # As per https://github.com/saschagrunert/kml-heatmap/issues/16
-            # Charterware coordinates are not at fixed intervals, making time/speed inference inaccurate
-
-            path_groups.append(current_path)
-            meta = _build_path_metadata_dict(
-                kml_file, current_path[0], airport_name, timestamp, end_timestamp
-            )
-            path_metadata.append(meta)
-
-        if element_coords > 0:
-            coord_type = (
-                "Point" if element_coords == 1 else f"Path ({element_coords} points)"
-            )
-            logger.debug(f"Element {idx}: {coord_type}")
-
-
-def _extract_gx_track_metadata(
-    placemarks: List[Any], namespaces: Dict[str, str], kml_file: str
-) -> Dict[str, Optional[str]]:
-    """
-    Extract metadata from gx:Track placemarks.
-
-    Args:
-        placemarks: List of Placemark XML elements
-        namespaces: XML namespace dict
-        kml_file: KML filename for ICAO extraction
-
-    Returns:
-        Dict with 'airport_name', 'timestamp', 'end_timestamp' keys
-    """
-    for placemark in placemarks:
-        # Check if this placemark contains gx:coord elements
-        placemark_gx_coords = find_xml_elements(
-            placemark, ".//gx:coord", ".//coord", namespaces
-        )
-
-        if placemark_gx_coords:
-            meta = extract_placemark_metadata(placemark, namespaces, kml_file)
-
-            if meta["timestamp"]:
-                logger.debug(f"Found gx:Track start timestamp: {meta['timestamp']}")
-            if meta["end_timestamp"]:
-                logger.debug(f"Found gx:Track end timestamp: {meta['end_timestamp']}")
-            if meta["timestamp"] is None and meta["airport_name"]:
-                logger.debug(
-                    f"No timestamp found for gx:Track with name: {meta['airport_name']}"
-                )
-
-            return meta
-
-    return {
-        "airport_name": None,
-        "timestamp": None,
-        "end_timestamp": None,
-        "year": None,
-    }
-
-
-def _extract_gx_when_elements(
-    placemarks: List[Any], gx_coords: List[Any], namespaces: Dict[str, str]
-) -> List[Any]:
-    """
-    Extract <when> elements that correspond to gx:coord elements.
-
-    Args:
-        placemarks: List of Placemark XML elements
-        gx_coords: List of gx:coord elements
-        namespaces: XML namespace dict
-
-    Returns:
-        List of <when> elements matching gx:coord count
-    """
-    for placemark in placemarks:
-        when_elems = find_xml_elements(placemark, ".//kml:when", ".//when", namespaces)
-        if when_elems and len(when_elems) == len(gx_coords):
-            return when_elems
-    return []
-
-
-def _parse_gx_coordinates(
-    gx_coords: List[Any],
-    when_elems: List[Any],
-    kml_file: str,
-    coordinates: List[List[float]],
-) -> List[List[Any]]:
-    """
-    Parse gx:coord elements into path coordinates.
-
-    Args:
-        gx_coords: List of gx:coord XML elements
-        when_elems: List of corresponding <when> timestamp elements
-        kml_file: Path to KML file (for error messages)
-        coordinates: Output list for [lat, lon] pairs
-
-    Returns:
-        List of parsed coordinates with altitude and timestamps
-    """
-    gx_path = []
-
-    for idx, gx_coord in enumerate(gx_coords):
-        if gx_coord.text is None:
-            continue
-
-        coord_text = gx_coord.text.strip()
-        if not coord_text:
-            continue
-
-        parts = coord_text.split()
-        if len(parts) < 2:
-            continue
-
-        try:
-            lon = float(parts[0])
-            lat = float(parts[1])
-            alt = float(parts[2]) if len(parts) >= 3 else None
-
-            # Use centralized validation and normalization
-            validated = validate_and_normalize_coordinate(
-                lat, lon, alt, f"{Path(kml_file).name} (gx:Track)"
-            )
-            if validated is None:
-                continue
-
-            lat, lon, alt = validated
-
-            # Get corresponding timestamp
-            timestamp_str = None
-            if idx < len(when_elems) and when_elems[idx].text:
-                timestamp_str = when_elems[idx].text.strip()
-
-            coordinates.append([lat, lon])
-
-            if alt is not None:
-                # Store as [lat, lon, alt, timestamp]
-                if timestamp_str:
-                    gx_path.append([lat, lon, alt, timestamp_str])
-                else:
-                    gx_path.append([lat, lon, alt])
-        except ValueError:
-            logger.debug(f"Failed to parse gx:coord: {coord_text}")
-            continue
-
-    return gx_path
-
-
-def process_gx_track(
-    gx_coords: List[Any],
-    placemarks: List[Any],
-    namespaces: Dict[str, str],
-    kml_file: str,
-    coordinates: List[List[float]],
-    path_groups: List[List[List[float]]],
-    path_metadata: List[Dict[str, Any]],
-) -> None:
-    """
-    Process Google Earth Track (gx:coord) elements.
-
-    Args:
-        gx_coords: List of gx:coord XML elements
-        placemarks: List of Placemark XML elements
-        namespaces: XML namespace dict
-        kml_file: Path to KML file
-        coordinates: Output list for all [lat, lon] pairs
-        path_groups: Output list for path groups with altitude
-        path_metadata: Output list for path metadata
-    """
-    if not gx_coords:
-        return
-
-    # Extract metadata from placemarks
-    track_meta = _extract_gx_track_metadata(placemarks, namespaces, kml_file)
-
-    # Extract timestamp elements
-    when_elems = _extract_gx_when_elements(placemarks, gx_coords, namespaces)
-
-    # Parse coordinates
-    gx_path = _parse_gx_coordinates(gx_coords, when_elems, kml_file, coordinates)
-
-    # Add gx:Track as a single path group
-    if gx_path:
-        path_groups.append(gx_path)
-        meta = _build_path_metadata_dict(
-            kml_file,
-            gx_path[0],
-            track_meta["airport_name"],
-            track_meta["timestamp"],
-            track_meta["end_timestamp"],
-        )
-        path_metadata.append(meta)
-
-    logger.debug(f"Parsed {len(gx_coords)} gx:coord elements into 1 track")
 
 
 def _parse_kml_tree(kml_file: str) -> Optional[Any]:
